+++ src/main/java/org/opensearch/knn/index/query/DefaultKNNWeight.java
@@ -15,7 +15,7 @@ import org.apache.lucene.util.Version;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.codec.util.KNNCodecUtil;
-import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 import org.opensearch.knn.index.memory.NativeMemoryAllocation;
 import org.opensearch.knn.index.memory.NativeMemoryCacheManager;
 import org.opensearch.knn.index.memory.NativeMemoryEntryContext;
@@ -51,7 +51,7 @@ public class DefaultKNNWeight extends KNNWeight {
         final SegmentReader reader,
         final FieldInfo fieldInfo,
         final SpaceType spaceType,
-        final KNNEngine knnEngine,
+        final VectorSearchEngine knnEngine,
         final VectorDataType vectorDataType,
         final byte[] quantizedVector,
         final float[] transformedVector,
@@ -182,7 +182,7 @@ public class DefaultKNNWeight extends KNNWeight {
         final SegmentReader reader,
         String cacheKey,
         final SpaceType spaceType,
-        final KNNEngine knnEngine,
+        final VectorSearchEngine knnEngine,
         final KNNQuery knnQuery,
         final VectorDataType vectorDataType,
         final byte[] quantizedVector,