+++ src/main/java/org/opensearch/knn/index/query/KNNWeight.java
@@ -35,6 +35,7 @@ import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 import org.opensearch.knn.index.quantizationservice.QuantizationService;
 import org.opensearch.knn.index.query.exactsearch.ExactSearcher;
 import org.opensearch.knn.index.query.exactsearch.ExactSearcher.ExactSearcherContext.ExactSearcherContextBuilder;
@@ -474,7 +475,7 @@ public abstract class KNNWeight extends Weight {
             return EMPTY_TOPDOCS;
         }
 
-        KNNEngine knnEngine;
+        VectorSearchEngine knnEngine;
         SpaceType spaceType;
         VectorDataType vectorDataType;
 
@@ -567,7 +568,7 @@ public abstract class KNNWeight extends Weight {
         final SegmentReader reader,
         final FieldInfo fieldInfo,
         final SpaceType spaceType,
-        final KNNEngine knnEngine,
+        final VectorSearchEngine knnEngine,
         final VectorDataType vectorDataType,
         final byte[] quantizedVector,
         final float[] transformedVector,
@@ -577,7 +578,7 @@ public abstract class KNNWeight extends Weight {
         final int k
     ) throws IOException;
 
-    protected void addExplainIfRequired(final KNNQueryResult[] results, final KNNEngine knnEngine, final SpaceType spaceType) {
+    protected void addExplainIfRequired(final KNNQueryResult[] results, final VectorSearchEngine knnEngine, final SpaceType spaceType) {
         if (knnQuery.isExplain()) {
             Arrays.stream(results).forEach(result -> {
                 if (KNNEngine.FAISS.getName().equals(knnEngine.getName()) && SpaceType.INNER_PRODUCT.equals(spaceType)) {
@@ -589,7 +590,7 @@ public abstract class KNNWeight extends Weight {
         }
     }
 
-    protected void addExplainIfRequired(final TopDocs results, final KNNEngine knnEngine, final SpaceType spaceType) {
+    protected void addExplainIfRequired(final TopDocs results, final VectorSearchEngine knnEngine, final SpaceType spaceType) {
         if (knnQuery.isExplain()) {
             Arrays.stream(results.scoreDocs).forEach(result -> {
                 if (KNNEngine.FAISS.getName().equals(knnEngine.getName()) && SpaceType.INNER_PRODUCT.equals(spaceType)) {
@@ -748,7 +749,7 @@ public abstract class KNNWeight extends Weight {
         if (fieldInfo == null) {
             return false;
         }
-        final KNNEngine knnEngine = FieldInfoExtractor.extractKNNEngine(fieldInfo);
+        final VectorSearchEngine knnEngine = FieldInfoExtractor.extractKNNEngine(fieldInfo);
         final List<String> engineFiles = KNNCodecUtil.getEngineFiles(
             knnEngine.getExtension(),
             knnQuery.getField(),