+++ src/main/java/org/opensearch/knn/plugin/stats/KNNStats.java
@@ -146,9 +146,9 @@ public class KNNStats {
     }
 
     private void addEngineStats(ImmutableMap.Builder<String, KNNStat<?>> builder) {
-        builder.put(StatNames.FAISS_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.FAISS)))
-            .put(StatNames.NMSLIB_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.NMSLIB)))
-            .put(StatNames.LUCENE_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.LUCENE)));
+        builder.put(StatNames.FAISS_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.FAISS::isInitialized)))
+            .put(StatNames.NMSLIB_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.NMSLIB::isInitialized)))
+            .put(StatNames.LUCENE_LOADED.getName(), createNodeStat(new LibraryInitializedSupplier(KNNEngine.LUCENE::isInitialized)));
     }
 
     private void addScriptStats(ImmutableMap.Builder<String, KNNStat<?>> builder) {