+++ src/test/java/org/opensearch/knn/index/codec/KNN990Codec/NativeEngines990KnnVectorsReaderTests.java
@@ -30,6 +30,7 @@ import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.codec.nativeindex.AbstractNativeEnginesKnnVectorsReader;
 import org.opensearch.knn.index.codec.KNNCodecTestUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
 import org.opensearch.knn.memoryoptsearch.VectorSearcher;
 import org.opensearch.knn.memoryoptsearch.VectorSearcherFactory;
@@ -90,7 +91,7 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
         FieldInfo[] fieldInfoArray = new FieldInfo[] { fieldInfo };
         final FieldInfos fieldInfos = new FieldInfos(fieldInfoArray);
 
-        KNNEngine mockFaiss = spy(KNNEngine.FAISS);
+        VectorSearchEngine mockFaiss = spy(KNNEngine.FAISS);
         VectorSearcherFactory mockFactory = mock(VectorSearcherFactory.class);
         VectorSearcher mockSearcher = mock(VectorSearcher.class);
         when(mockSearcher.getByteVectorValues(any())).thenReturn(mock(ByteVectorValues.class));
@@ -131,7 +132,7 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
 
     @SneakyThrows
     public void testWhenMemoryOptimizedSearchIsEnabled_mixedCase() {
-        KNNEngine mockFaiss = spy(KNNEngine.FAISS);
+        VectorSearchEngine mockFaiss = spy(KNNEngine.FAISS);
         VectorSearcherFactory mockFactory = mock(VectorSearcherFactory.class);
         when(mockFactory.createVectorSearcher(any(), any(), any(), any(), any())).thenReturn(mock(VectorSearcher.class));
         when(mockFaiss.getVectorSearcherFactory()).thenReturn(mockFactory);
@@ -187,7 +188,7 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
         }
     }
 
-    private static FieldInfo createFieldInfo(final String fieldName, final KNNEngine engine, final int fieldNo) {
+    private static FieldInfo createFieldInfo(final String fieldName, final VectorSearchEngine engine, final int fieldNo) {
         final KNNCodecTestUtil.FieldInfoBuilder builder = KNNCodecTestUtil.FieldInfoBuilder.builder(fieldName);
         builder.fieldNumber(fieldNo);
         if (engine != null) {