+++ src/test/java/org/opensearch/knn/integ/BinaryIndexIT.java
@@ -23,6 +23,7 @@ import org.opensearch.knn.TestUtils;
 import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -43,9 +44,9 @@ public class BinaryIndexIT extends KNNRestTestCase {
     private static TestUtils.TestData testData;
     private static final int NEVER_BUILD_GRAPH = -1;
     private static final int ALWAYS_BUILD_GRAPH = 0;
-    private final KNNEngine engine;
+    private final VectorSearchEngine engine;
 
-    public BinaryIndexIT(KNNEngine engine) {
+    public BinaryIndexIT(VectorSearchEngine engine) {
         this.engine = engine;
     }
 
@@ -236,7 +237,7 @@ public class BinaryIndexIT extends KNNRestTestCase {
     }
 
     private void createKnnHnswBinaryIndex(
-        final KNNEngine knnEngine,
+        final VectorSearchEngine knnEngine,
         final String indexName,
         final String fieldName,
         final int dimension,
@@ -257,8 +258,12 @@ public class BinaryIndexIT extends KNNRestTestCase {
         createKnnIndex(indexName, buildKNNIndexSettings(threshold), knnIndexMapping);
     }
 
-    private void createKnnHnswBinaryIndex(final KNNEngine knnEngine, final String indexName, final String fieldName, final int dimension)
-        throws IOException {
+    private void createKnnHnswBinaryIndex(
+        final VectorSearchEngine knnEngine,
+        final String indexName,
+        final String fieldName,
+        final int dimension
+    ) throws IOException {
         createKnnHnswBinaryIndex(knnEngine, indexName, fieldName, dimension, ALWAYS_BUILD_GRAPH);
     }
 }